requirements.txt
yt-dlp
ffmpeg-python
openai-whisper
langchain
langchain-core
langchain-google-genai
langchain-community
langgraph
sentence-transformers
pydantic
fastapi
uvicorn
python-dotenv
requests
